+++ index.html
@@ -22,8 +22,8 @@
                 <h2>Hello~!</h2>
                 <div>
                     <p>site navigation will require client-side JS, please enable it.</p>
-                    <p>this site also requires the browser has at least been updated beyond 2014</p>
-                    <a href="/room/">got it..</a>
+                    <p>this site also requires the browser has at least been updated beyond 2017</p>
+                    <a href="room/#">got it..</a>
                 </div>
             </div>
         </div>
