+++ room/index.html
@@ -16,6 +16,7 @@
 	<meta name="viewport" content="width=device-width, initial-scale=0.6">
 
 	<script src="/shared/global.js"></script>
+	<script type="module" src="/room/room.js"></script>
 </head>
 
 <body class="umbrella-background">
@@ -23,7 +24,7 @@
 		<div class="main-container big-box">
 			
 			<div class="max-item home-content">
-				<main data-route-vanilla-ele="router-wrap" class="max-item main-content">
+				<main data-vanilla-route-ele="router-wrap" class="max-item main-content">
 					<!-- USE HASH ROUTING FOR THIS PAGE -->
 				</main>
 				<nav class="side-nav">
@@ -34,8 +35,8 @@
 						<div class="spacer"></div>
 						<div class="divider-vertical"></div>
 						<div class="side-nav-content">
-							<a data-vanilla-route-link="spa" href="/room/">room</a>
-							<a data-vanilla-route-link="spa" href="/about/">about</a>
+							<a data-vanilla-route-link="spa" href="/">room</a>
+							<a data-vanilla-route-link="spa" href="/about">about</a>
 						</div>
 
 					</div>
